+++ src/processor.rs
@@ -15,19 +15,11 @@ pub type Timestamp = DateTime<Utc>;
 pub trait Processor: Send + 'static {
   /// Process a log line. The line is guaranteed to be non-empty and does not contain a newline character.
   ///
-  /// Return `true` if the line is written to the underlying output stream (maybe an empty line, maybe need flush).
-  /// Return `false` if the line is ignored.
-  ///
   /// The line may be an [EMF](https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/CloudWatch_Embedded_Metric_Format_Specification.html) line.
   /// You can use the [`is_emf`](crate::is_emf) util function to check if it is.
-  fn process(&mut self, line: String, timestamp: Timestamp) -> impl Future<Output = bool> + Send;
+  fn process(&mut self, line: String, timestamp: Timestamp) -> impl Future<Output = ()> + Send;
 
   /// Flushes the underlying output stream, ensuring that all intermediately buffered contents reach their destination.
-  fn flush(&mut self) -> impl Future<Output = ()> + Send;
-
   /// This method is called when the current invocation is about to end (the proxy got the request of `invocation/next`).
-  fn truncate(&mut self) -> impl Future<Output = ()> + Send {
-    // do nothing by default
-    async {}
-  }
+  fn truncate(&mut self) -> impl Future<Output = ()> + Send;
 }